backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.1.8
cryptography==46.0.6
dnspython==2.7.0
ecdsa==0.19.2
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.128.8
fpdf==1.7.2
h11==0.16.0
idna==3.11
joblib==1.5.3
lxml==6.0.2
nltk==3.9.2
numpy
passlib==1.7.4
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==11.3.0
pyasn1==0.6.3
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pypdfium2==5.6.0
python-docx==1.2.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
regex==2026.1.15
requests==2.32.5
rsa==4.9.1
sentence-transformers
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.49.3
textblob==0.19.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.39.0
faiss-cpu